clusterisation, inpainting done. trash files deleted. todo: superresolution. note: checkpoints folder deleted caused by its size
